+++ lib/widgets/universal.dart
@@ -975,7 +975,7 @@ class SimpleButton extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final Widget current = child ??
-        BasisText(text,
+        BText(text,
             textAlign: TextAlign.start,
             style: textStyle,
             maxLines: maxLines,